# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: JOE Hub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quizzes:
    get:
      operationId: listQuizzes
      tags: [quizzes]
      summary: List quizzes
      parameters:
        - name: courseId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of quizzes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quiz"
    post:
      operationId: createQuiz
      tags: [quizzes]
      summary: Create a quiz (admin/mentor)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
  /quizzes/{id}:
    get:
      operationId: getQuiz
      tags: [quizzes]
      summary: Get a quiz (question options omit isCorrect for students)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
        "404":
          description: Not found
    patch:
      operationId: updateQuiz
      tags: [quizzes]
      summary: Update a quiz (admin/mentor)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
        "404":
          description: Not found
    delete:
      operationId: deleteQuiz
      tags: [quizzes]
      summary: Delete a quiz (admin/mentor)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /quizzes/{id}/questions:
    get:
      operationId: listQuizQuestions
      tags: [quizzes]
      summary: List questions of a quiz (admin/mentor view, includes correct answers)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
    post:
      operationId: createQuestion
      tags: [quizzes]
      summary: Create a question with options for a quiz (admin/mentor)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /questions/{id}:
    patch:
      operationId: updateQuestion
      tags: [quizzes]
      summary: Update a question and replace its options (admin/mentor)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "404":
          description: Not found
    delete:
      operationId: deleteQuestion
      tags: [quizzes]
      summary: Delete a question (admin/mentor)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /quizzes/{id}/attempts:
    post:
      operationId: startQuizAttempt
      tags: [quizzes]
      summary: Start a new timed attempt on a quiz as the current student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Created attempt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attempt"
  /quizzes/{id}/attempts/me:
    get:
      operationId: listMyQuizAttempts
      tags: [quizzes]
      summary: List the current student's attempt history for a quiz
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of attempts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attempt"
  /attempts/{id}:
    get:
      operationId: getAttempt
      tags: [quizzes]
      summary: Get an attempt, including answered questions so far
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Attempt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attempt"
        "404":
          description: Not found
  /attempts/{id}/answers:
    post:
      operationId: submitAnswer
      tags: [quizzes]
      summary: Submit (or change) an answer for one question in an in-progress attempt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnswerInput"
      responses:
        "200":
          description: Recorded answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Answer"
  /attempts/{id}/submit:
    post:
      operationId: submitQuizAttempt
      tags: [quizzes]
      summary: Finalize an attempt and automatically grade it
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Graded attempt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attempt"

  # ---------- Announcements ----------
...
components:
  schemas:
...
    AttemptStatus:
      type: string
      enum: [in_progress, completed]
...
    Quiz:
      type: object
      required: [id, courseId, title, questionCount, createdAt]
      properties:
        id: { type: integer }
        courseId: { type: integer }
        moduleId: { type: ["integer", "null"] }
        title: { type: string }
        description: { type: ["string", "null"] }
        timeLimitMinutes: { type: ["integer", "null"] }
        questionCount: { type: integer }
        createdAt: { type: string, format: date-time }
    QuizInput:
      type: object
      required: [courseId, title]
      properties:
        courseId: { type: integer }
        moduleId: { type: integer }
        title: { type: string, minLength: 1 }
        description: { type: string }
        timeLimitMinutes: { type: integer }
    QuizUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        timeLimitMinutes: { type: integer }
    Option:
      type: object
      required: [id, questionId, optionText, order]
      properties:
        id: { type: integer }
        questionId: { type: integer }
        optionText: { type: string }
        isCorrect: { type: boolean }
        order: { type: integer }
    OptionInput:
      type: object
      required: [optionText, isCorrect]
      properties:
        optionText: { type: string, minLength: 1 }
        isCorrect: { type: boolean }
        order: { type: integer, default: 0 }
    Question:
      type: object
      required: [id, quizId, question, order, options]
      properties:
        id: { type: integer }
        quizId: { type: integer }
        question: { type: string }
        order: { type: integer }
        options:
          type: array
          items:
            $ref: "#/components/schemas/Option"
    QuestionInput:
      type: object
      required: [question, options]
      properties:
        question: { type: string, minLength: 1 }
        order: { type: integer, default: 0 }
        options:
          type: array
          minItems: 2
          items:
            $ref: "#/components/schemas/OptionInput"
    QuestionUpdate:
      type: object
      properties:
        question: { type: string, minLength: 1 }
        order: { type: integer }
        options:
          type: array
          minItems: 2
          items:
            $ref: "#/components/schemas/OptionInput"
    Answer:
      type: object
      required: [id, attemptId, questionId, selectedOptionId, isCorrect]
      properties:
        id: { type: integer }
        attemptId: { type: integer }
        questionId: { type: integer }
        selectedOptionId: { type: integer }
        isCorrect: { type: boolean }
    AnswerInput:
      type: object
      required: [questionId, selectedOptionId]
      properties:
        questionId: { type: integer }
        selectedOptionId: { type: integer }
    Attempt:
      type: object
      required: [id, quizId, studentId, totalQuestions, status, startedAt]
      properties:
        id: { type: integer }
        quizId: { type: integer }
        studentId: { type: integer }
        score: { type: ["integer", "null"] }
        totalQuestions: { type: integer }
        status: { $ref: "#/components/schemas/AttemptStatus" }
        startedAt: { type: string, format: date-time }
        submittedAt: { type: ["string", "null"], format: date-time }
        answers:
          type: array
          items:
            $ref: "#/components/schemas/Answer"
...